Update patch set 1
Patch Set 1:

I used the "Cherry Pick To" button in the gerrit interface and told it to cherry pick to the stable branch.

The other way of doing it would have been to remove the existing Change-ID from the commit message so that a new one is generated. Change-Ids are used as unique identifier in Gerrit, so to avoid duplication you need a new one. (The Cherry-pick button gets around that, though).

Patch-set: 1